Fix x509 vs root store comparisons (#151)
The root store doesn't keep the x509 padding, so we were not correctly finding the root to validate CRL signatures. Thus, we need to add the padding back. The first commit did this directly in the checks, but I thought it was ugly, so I've added the TrustAnchorView which lazily adds the padding.

The important test is trust_anchor_view_provides_compatibility_with_x509 which creates a cert, adds it to the root store, and ensure we can successfully compare.

This fixes the Python test test_should_connect_and_select_with_crl_enabled.